# api/openapi/api-gateway.yaml
openapi: 3.1.0
info:
  title: Personal Finance OS API Gateway
  version: 0.4.0
servers:
  - url: http://localhost:8080
paths:
...
  /imports/raw:
    post:
      summary: Upload a raw statement import into MongoDB and enqueue parse work
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Raw import accepted
  /imports/{import_id}:
    get:
      summary: Get raw import status from MongoDB
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: import_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Raw import metadata
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...